split the AQ code into separate layes to make the listener approach easier, better log output on completion and thread tidying up
