+++ IoTDBJDBC/src/main/java/com/oracle/demo/timg/iot/iotdbjdbc/aqdata/NormalizedData.java
@@ -13,6 +13,7 @@ public class NormalizedData {
 	private String digitalTwinInstanceId;
 	private String contentPath;
 	private String timeObserved;
+	private String contentType;
 	private String content;
 	public static final String SQL_QUEUE_NAME = "normalized_data";
 }
